+++ javaagent-tooling/src/main/java/io/opentelemetry/javaagent/tooling/ignore/AdditionalLibraryIgnoredTypesConfigurer.java
@@ -273,7 +273,6 @@ public class AdditionalLibraryIgnoredTypesConfigurer implements IgnoredTypesConf
     // kotlin, note we do not ignore kotlinx because we instrument coroutines code
     builder
         .ignoreClass("kotlin.")
-        .allowClass("kotlin.coroutines.jvm.internal.DebugProbesKt")
         .allowClass("kotlin.coroutines.jvm.internal.CompletedContinuation");
   }
 }
